Haskell: More type safety for keys
Since we have a few maps, all addressed by keys of the same type
(String), there is a risk of introducing bugs by using the wrong key,
e.g. confusing the key for a block with that of a rule. This uses the
type system to prevent that, and allows for nicer type signatures. Yay!