Switch to unbound to handle names
and make unification variables unique using the unbound machinery.

Also, in logic specifications, be explicit about what is all-quantified
in the rules.

Some fragments of quanitifiers are there, but not done yet.

This commit was brought to you by a 1½h delay of Condor airlines, in
cooperation with Vancouver Airport Wireless.