Localize the proof machine
using i18next as suggested by @NicoleRauch. This fixes #23.

I did not yet translate the #help dialog. It seems it would make sense
to define in this element two divs, one with all the English content and
one with all the German content, and show only the right one. This does
not seem to be directly supported with i18next.