Refactor prepare
to return a more polished and selfcontained description of the proof,
with renaming and scopes taken care of. This simplifies later stages
(look, no "task" any more), and also prepares for #24.

I removed one of the test cases; the test suite looks a bit too messy
anyway right now. At least I could not make quick sense of what belongs
to what :-). We should rely more on the tests based on the examples
directory, that is easier to maintain.