Get rid of LabelConnections
we now return a map from PortSpec to Proposition (more preparation for
bug #24) and a simple enumeration type for the status of each
connection.

This patch is incomplete, as there is no nice way to represent the map
from PortSpec in JSON. So I’ll first refacort PortSpec into a simple
product type, and then work on this can continue.