Implement rule export including local hypotheses
This fixes #11; open new tickets for any follow up issues.

It fails (rightfully) if it finds no possible target for a given local
hypothesis. If there are multiple possible targets, it chooses one
arbitrarily. This is suboptimal, but we have to see how important it is
in practice.

Thanks to @gattschardo for most of the workon this issue.