New graph view on a proof
This new (and yet unused) module creates a very nice graph structure
from a proof. Our graphs are 4-partite, when considering blocks,
connections, in-ports and out-ports as nodes. This module statically
types this. Furthermore, we do not change our graphs, so it makes sense
to “tie the know” and get rid of maps when traversing the graph: By
having successors and predecessors point to the corresponding node
object in heap, the pointer structure on the heap is precisely the graph
we want. This should make traversals quick.

If we need the extra performance boost we could add an Int field to
Node, allowing a traversal to use an IntMap or IntSet to keep track of
data or flags per node.